requirements.txt
python-dotenv>=1.0.0
paramiko>=3.0.0
python-gnupg>=0.5.0
redis>=5.0.0
rq>=1.15.0
openai>=1.0.0
flask>=3.0.0
